vision_segmentation/requirements.txt
datasets>=2.0
torch>=2.0
torchvision
numpy
Pillow
matplotlib
tqdm
opencv-python-headless
albumentations
